[#156] Move families to 'Relude.Extra.Type'
