+++ src/Relude/Extra/Type.hs
@@ -1,18 +1,27 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE CPP                 #-}
+{-# LANGUAGE DataKinds           #-}
+{-# LANGUAGE PolyKinds           #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications    #-}
+{-# LANGUAGE TypeFamilies        #-}
+{-# LANGUAGE TypeOperators       #-}
 
 {- |
 Copyright:  (c) 2018-2019 Kowainik
+                2019 TheMatten
 License:    MIT
 Maintainer: Kowainik <xrom.xkov@gmail.com>
 
-Contains useful functions to work with Types.
+Contains useful utilities to work with Types.
 -}
 
 module Relude.Extra.Type
        ( typeName
+       , type (++)
+       , AllHave
+       , Fst
+       , Snd
        ) where
 
 import Relude
@@ -22,6 +31,9 @@ import Type.Reflection (Typeable, typeRep)
 #else
 import Data.Typeable (Typeable, typeRep)
 #endif
+
+-- $setup
+-- >>> :set -XDataKinds -XTypeOperators
 
 {- | Gets a string representation of a type.
 
@@ -43,3 +55,63 @@ typeName = show (typeRep @a)
 typeName = show (typeRep (Proxy @a))
 #endif
 {-# INLINE typeName #-}
+
+infixr 5 ++
+{- | Concatenates type-level lists.
+
+>>> :kind! '[ 'Just 5, 'Nothing] ++ '[ 'Just 3, 'Nothing, 'Just 1]
+'[ 'Just 5, 'Nothing] ++ '[ 'Just 3, 'Nothing, 'Just 1] :: [Maybe
+                                                              Nat]
+= '[ 'Just 5, 'Nothing, 'Just 3, 'Nothing, 'Just 1]
+>>> :kind! '[] ++ '[ 'Just 3, 'Nothing, 'Just 1]
+'[] ++ '[ 'Just 3, 'Nothing, 'Just 1] :: [Maybe Nat]
+= '[ 'Just 3, 'Nothing, 'Just 1]
+-}
+type family (++) (xs :: [k]) (ys :: [k]) :: [k] where
+    '[]       ++ ys = ys
+    (x ': xs) ++ ys = x ': xs ++ ys
+
+{- | Builds combined 'Constraint' by applying Constraint constructor to all
+elements of type-level list.
+
+>>> :kind! AllHave Show '[Int, Text, Double]
+AllHave Show '[Int, Text, Double] :: Constraint
+= (Show Int, (Show Text, (Show Double, () :: Constraint)))
+
+which is equivalent to:
+
+@
+(Show Int, Show Text, Show Double) :: Constraint
+@
+-}
+type family AllHave (f :: k -> Constraint) (xs :: [k]) :: Constraint where
+    AllHave _ '[]       = ()
+    AllHave f (x ': xs) = (f x, AllHave f xs)
+
+{- | Returns first element of tuple type (with kind @*@) or type-level tuple
+(with kind @(k1, k2)@, marked by prefix quote).
+
+>>> :kind! Maybe (Fst '(Int, Text))
+Maybe (Fst '(Int, Text)) :: *
+= Maybe Int
+>>> :kind! Maybe (Fst (Int, Text))
+Maybe (Fst (Int, Text)) :: *
+= Maybe Int
+-}
+type family Fst (t :: k) :: k' where
+    Fst '(x, _) = x
+    Fst  (x, _) = x
+
+{- | Returns second element of tuple type (with kind @*@) or type-level tuple
+(with kind @(k1, k2)@, marked by prefix quote).
+
+>>> :kind! Maybe (Snd '(Int, Text))
+Maybe (Snd '(Int, Text)) :: *
+= Maybe Text
+>>> :kind! Maybe (Snd (Int, Text))
+Maybe (Snd (Int, Text)) :: *
+= Maybe Text
+-}
+type family Snd (t :: k) :: k' where
+    Snd '(_, y) = y
+    Snd  (_, y) = y
